feat(build): automate legacy *.icns icon creation
Create the legacy `*.icns` directly from a full resolution single PNG, 
and automate it via Make targets.

This replaces the old Affinity Designer document which exported tons of 
separate PNG images as various resolutions in `*.iconset` folders, which
we then had to manually convert to a `.icns` file with `iconutil`.